// src/zed-theme.d.ts
export type AppearanceContent = 'light' | 'dark';
export type AccentContent = string | null;
/**
 * The background appearance of the window.
 */
export type WindowBackgroundContent = 'opaque' | 'transparent' | 'blurred';
export type FontStyleContent = 'normal' | 'italic' | 'oblique';

/**
 * The content of a serialized theme family.
 */
export interface ThemeFamilyContent {
  author: string;
  name: string;
  themes: ThemeContent[];
  [k: string]: unknown;
}
/**
 * The content of a serialized theme.
 */
export interface ThemeContent {
  appearance: AppearanceContent;
  name: string;
  style: ThemeStyleContent;
  [k: string]: unknown;
}